requirements.txt
Django>=5.0
djangorestframework>=3.14
psycopg[binary]>=3.1
celery>=5.3
redis>=5.0
python-decouple>=3.8
reportlab>=4.0
